+++ test/EFieldTest.py
@@ -93,9 +93,9 @@ class EFieldTest(unittest.TestCase):
         self._sc = Checker(self)
 
         try:
-            self.__seed  = long(binascii.hexlify(os.urandom(16)), 16)
+            self.__seed  = int(binascii.hexlify(os.urandom(16)), 16)
         except NotImplementedError:
-            self.__seed  = long(time.time() * 256) 
+            self.__seed  = int(time.time() * 256) 
 #        self.__seed =241361343400098333007607831038323262554
             
         self.__rnd = random.Random(self.__seed)
