+++ test/NTPTest.py
@@ -84,8 +84,8 @@ class NTPTest(unittest.TestCase):
                       "NX_DATE_TIME":"string", "ISO8601":"string", "NX_CHAR":"string", "NX_BOOLEAN":"bool"}
 
         ## map of type : converting function
-        self._convert = {"float16":float,"float32":float, "float64":float, "float":float, "int64":long, "int32":int,  
-                         "int16":int, "int8":int, "int":int, "uint64":long, "uint32":long, "uint16":int,
+        self._convert = {"float16":float,"float32":float, "float64":float, "float":float, "int64":int, "int32":int,  
+                         "int16":int, "int8":int, "int":int, "uint64":int, "uint32":int, "uint16":int,
                          "uint8":int, "uint":int, "string":str, "bool":Converters.toBool}
         
         ## map of tag attribute types 
@@ -104,10 +104,10 @@ class NTPTest(unittest.TestCase):
 
 
         try:
-            self.__seed  = long(binascii.hexlify(os.urandom(16)), 16)
+            self.__seed  = int(binascii.hexlify(os.urandom(16)), 16)
         except NotImplementedError:
             import time
-            self.__seed  = long(time.time() * 256) # use fractional seconds
+            self.__seed  = int(time.time() * 256) # use fractional seconds
          
         self.__rnd = random.Random(self.__seed)
