+++ test/TangoFieldTagWriterTest.py
@@ -62,10 +62,10 @@ class TangoFieldTagWriterTest(unittest.TestCase):
 
         try:
             ## random seed
-            self.seed  = long(binascii.hexlify(os.urandom(16)), 16)
+            self.seed  = int(binascii.hexlify(os.urandom(16)), 16)
         except NotImplementedError:
             import time
-            self.seed  = long(time.time() * 256) # use fractional seconds
+            self.seed  = int(time.time() * 256) # use fractional seconds
 
         self.__rnd = random.Random(self.seed)
