+++ test/TgDeviceTest.py
@@ -50,7 +50,6 @@ from nxswriter.DataSourcePool import DataSourcePool
 from nxswriter import DataSources
 
 import threading
-import thread
 
 ## if 64-bit machione
 IS64BIT = (struct.calcsize("P") == 8)
@@ -89,10 +88,10 @@ class TgDeviceTest(unittest.TestCase):
         self._bfloat = "float64" if IS64BIT else "float32"
 
         try:
-            self.__seed  = long(binascii.hexlify(os.urandom(16)), 16)
+            self.__seed  = int(binascii.hexlify(os.urandom(16)), 16)
         except NotImplementedError:
             import time
-            self.__seed  = long(time.time() * 256) # use fractional seconds
+            self.__seed  = int(time.time() * 256) # use fractional seconds
          
         self.__rnd = random.Random(self.__seed)
