+++ test/ThreadPoolTest.py
@@ -27,7 +27,7 @@ import random
 import struct
 import binascii
 import time
-import Queue
+import queue
 import json
 
 from nxswriter.ThreadPool import ThreadPool
@@ -153,10 +153,10 @@ class ThreadPoolTest(unittest.TestCase):
         self._bfloat = "float64" if IS64BIT else "float32"
 
         try:
-            self.__seed  = long(binascii.hexlify(os.urandom(16)), 16)
+            self.__seed  = int(binascii.hexlify(os.urandom(16)), 16)
         except NotImplementedError:
             import time
-            self.__seed  = long(time.time() * 256) # use fractional seconds
+            self.__seed  = int(time.time() * 256) # use fractional seconds
          
         self.__rnd = random.Random(self.__seed)
