+++ test/UINT32decoderTest.py
@@ -60,9 +60,9 @@ class UINT32decoderTest(unittest.TestCase):
 
 
         try:
-            self.__seed  = long(binascii.hexlify(os.urandom(16)), 16)
+            self.__seed  = int(binascii.hexlify(os.urandom(16)), 16)
         except NotImplementedError:
-            self.__seed  = long(time.time() * 256) 
+            self.__seed  = int(time.time() * 256) 
          
         self.__rnd = random.Random(self.__seed)
